+++ openharmony/arkcompiler_runtime_core/static_core/runtime/mem/refstorage/reference_storage.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021-2024 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -84,7 +84,7 @@ public:
      * @param capacity minimum number of references for this frame
      * @return true if current frame can store at least `size` references, false otherwise.
      */
-    bool EnsureLocalCapacity(size_t capacity);
+    bool EnsureLocalCapacity(uint32_t capacity);
 
     /// Get all objects in global & local storage. Use for debugging only
     PandaVector<ObjectHeader *> GetAllObjects();
